server/requirements.txt
corsheaders
psycopg2
Django
djangorestframework
pandas 
librosa
absl-py==2.1.0
astunparse==1.6.3
audioread==3.0.1
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==5.1.1
Flask==3.0.3
Flask-Cors==4.0.0
flatbuffers==24.3.25
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.2
gunicorn==22.0.0
h5py==3.11.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.4.0
keras==3.3.2
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.1
llvmlite==0.42.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
msgpack==1.0.8
namex==0.0.8
numba==0.59.1
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
platformdirs==4.2.1
pooch==1.8.1
protobuf==4.25.3
pycparser==2.22
Pygments==2.17.2
requests==2.31.0
rich==13.7.1
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
soundfile==0.12.1
soxr==0.3.7
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.4.0
typing_extensions==4.11.0
urllib3==2.2.1
waitress==3.0.0
Werkzeug==3.0.2
wrapt==1.16.0
